[FIX] 오류 제거, 폴더명 변경, 미사용 파일 제거, gitignore에 saved추가
