+++ baseline/detection.py
